updated code using components strategy
